404 not found added

+++ src/main/java/edu/lab/back/db/dao/implementations/postgres/BaseCrudDaoImpl.java
@@ -34,6 +34,10 @@ public abstract class BaseCrudDaoImpl<EntityType, IdType> implements BaseCrudDao
     @Override
     public EntityType deleteById(final IdType id, Class<EntityType> entityClass) {
         final EntityType city = this.getById(id, entityClass);
+        if (city == null) {
+            return null;
+        }
+
         this.getEntityManager().remove(city);
         return city;
     }
